+++ src/test/java/com/ejoongseok/wmslive/common/ApiTest.java
@@ -1,0 +1,19 @@
+package com.ejoongseok.wmslive.common;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class ApiTest {
+    @LocalServerPort
+    private int port;
+
+    @BeforeEach
+    void setUp() {
+        if (RestAssured.UNDEFINED_PORT == RestAssured.port) {
+            RestAssured.port = port;
+        }
+    }
+}
