+++ src/test/java/com/ejoongseok/wmslive/product/feature/RegisterProductTest.java
@@ -1,33 +1,21 @@
 package com.ejoongseok.wmslive.product.feature;
 
+import com.ejoongseok.wmslive.common.ApiTest;
 import com.ejoongseok.wmslive.product.domain.Category;
 import com.ejoongseok.wmslive.product.domain.ProductRepository;
 import com.ejoongseok.wmslive.product.domain.TemperatureZone;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class RegisterProductTest {
-    @LocalServerPort
-    private int port;
+class RegisterProductTest extends ApiTest {
     @Autowired
     private ProductRepository productRepository;
-
-    @BeforeEach
-    void setUp() {
-        if (RestAssured.UNDEFINED_PORT == RestAssured.port) {
-            RestAssured.port = port;
-        }
-    }
 
     @Test
     @DisplayName("상품을 등록한다.")
